Topology_Project/requirements.txt
openai>=1.0.0
groq>=0.5.0
pydantic>=2.0
pytest>=7.0
matplotlib>=3.0
tiktoken>=0.5
python-dotenv>=1.0